(new) 帳簿分別用のタグのREST API作成 issue #7 author: jagasorutomatai description: 	帳簿分別用のタグのREST API作成しました。
