+++ apiv1/serializers.py
@@ -1,9 +1,16 @@
-from household.models import Book
+from household.models.book import Book
+from household.models.tag import Tag
 from rest_framework import serializers
 
 
 class BookSerializer(serializers.ModelSerializer):
     class Meta:
         model = Book
-        fields = ['id', 'title', 'money', 'date',
+        fields = ['id', 'title', 'money', 'date', 'tag',
                   'account', 'created_at', 'updated_at']
+
+
+class TagSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tag
+        fields = ['id', 'name', 'created_at', 'updated_at']
