+++ apiv1/urls.py
@@ -4,7 +4,8 @@ from rest_framework import routers
 from . import views
 
 router = routers.DefaultRouter()
-router.register('books', views.BookViewSet)
+router.register('household/books', views.BookViewSet)
+router.register('household/tags', views.TagViewSet)
 
 app_name = 'apiv1'
 urlpatterns = [
