+++ household/models/__init__.py
@@ -1,0 +1,2 @@
+from household.models.book import Book  # NOQA
+from household.models.tag import Tag  # NOQA
